Provisioner now instantiates the heartbeat.
I have NO idea why this isn't actually used by the kernel, but as far as
I can tell it ain't. So this spins up a zmqstream and says "Ping" at it.
The jank is that unlike the Java version, I gotta use the stream here,
which weirdly is harder since I got specifically used to the more direct
version in Ignition. At any rate, if either the kernel or the server dies
the other will see that and close.

I'll finally note that this probably isn't correct as a process. I dunno
how to use await. We need to hear the reply from the heartbeat - that's
what it _does_ - so the stream version of the socket in Python uses a
callback set by on_recv. That's setting the last_heartbeat, and each
poll the function just sends a ping and checks if the heartbeat implies
cardiac arrest. It'll be approximately one beat behind (or basically lub
before the last dub), but it'll on the whole be fine so long as timedelta
is set sufficiently larger than the beat rate (which seems about every 3
seconds?)